Test and fix run_env_hooks
The quoting around the ${hook} variable was causing both hook file and
function root to be globbed as a single arg (is that the right use of "glob"?
who knows).

Unquoting causes a shellcheck error which isn't ideal, will investigate
whether there's a more legitimate way to get this interpreted as two
arguments.